Handle active wallet by env variable and cleanups
