+++ scripts/run/distribute_gauge_rewards.py
@@ -2,10 +2,8 @@
 from brownie import accounts
 
 from scripts import contracts
-from scripts.helper import get_addresses, gas_strategy
+from scripts.helper import get_addresses, gas_strategy, DEPLOY_ACCT
 
-# DEPLOY_ACCT = accounts.load('hardhat')
-DEPLOY_ACCT = accounts.load('deployve')
 
 addresses = get_addresses()
