+++ scripts/run/poke_distributor_mining_parameters.py
@@ -1,10 +1,6 @@
 #!/usr/bin/env python
-from brownie import accounts
-
 from scripts import contracts
-from scripts.helper import get_addresses, gas_strategy
-
-DEPLOY_ACCT = accounts.load('deployve')
+from scripts.helper import get_addresses, gas_strategy, DEPLOY_ACCT
 
 addresses = get_addresses()
