+++ scripts/run/toggle_gauge_distributor.py
@@ -1,14 +1,10 @@
 #!/usr/bin/env python
-from brownie import accounts
 import json
 import time
 
 from scripts import contracts
-from scripts.helper import get_addresses, gas_strategy
+from scripts.helper import get_addresses, gas_strategy, DEPLOY_ACCT
 
-
-DEPLOY_ACCT = accounts.load('hardhat')
-# DEPLOY_ACCT = accounts.load('deployve')
 
 addresses = get_addresses()
